requirements.txt
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
anthropic==0.21.3
openai==1.14.3
python-dotenv==1.0.1
requests==2.31.0
sendgrid==6.11.0
pytz==2024.1
httpx==0.27.0
